Fixed np.int in iterative_recon_alg.py
Using the Python installation of TIGRE raised the error: 
AttributeError: module 'numpy' has no attribute 'int'.
`np.int` was a deprecated alias for the builtin `int`.
This happens while running demos and functions utilizing iterative_recon_alg. 
changing `np.int` to `np.int32` fixes the issues, as proposed by numpy itself.

+++ Python/tigre/algorithms/iterative_recon_alg.py
@@ -380,7 +380,7 @@ class IterativeReconAlg(object):
         :return: None
         """
 
-        ang_index = self.angle_index[iteration].astype(np.int)
+        ang_index = self.angle_index[iteration].astype(np.int32)
 
         self.res += (
             self.lmbda
